requirements.txt
# requirements.txt
aiomysql==0.2.0
httpx==0.23.3
fastapi==0.95.0
uvicorn==0.21.1
python-dotenv==1.0.0
aiofiles==23.1.0
pydantic==1.10.7
websockets==10.4
python-multipart==0.0.6
redis==4.5.4